Fixed Case 3
userAddMovie returns String array. Needs to be passed into addMovie. userAddMovie now has a while loop to ensure that a name is entered.
addMovie currently does not increase node count or add to the list. May need to create an OrderedLInkedList that contains additional method for inserting into the list.